templates/model_products/main_product/images/byoc/requirements.txt
pandas==2.3.1
tensorflow==2.19.0
pyyaml==6.0.2
google-cloud-storage==3.4.1
google-cloud-bigquery==3.36.0
db-dtypes==1.4.3
pyarrow==19.0.0
